requirements.txt
pandas
numpy
sentence-transformers
pinecone-client
python-dotenv
openai
tqdm
